Add files via upload
Chapter 3 of Thesis manuscript : Forecasting yearly record probabilities from historical measurements
This is the last version of the code. To run it one just need as input the files "region"_observations.rds and "region"_information.rds. This file produces all figures in chapter 3 of the Thesis manuscript : 
1. Map of France with location of stations (Figure 3.1)
2. time-series panel (Figure 3.2)
3. Schema of theoretical curves (Figure 3.3)
4. record probability (Figure 3.4)
5.  record probability ratio (Figure 3.5)
5. Pmodel/Pindep (Figure 3.6)

In addition, this file contains : 

6. Beta QQ-plot
7. simulations